Migrate Multicolor ColorMode
